\name{PartialCox-package}
\alias{PartialCox-package}
\alias{PartialCox}
\docType{package}
\title{
Partial Cox regression model training and application
}
\description{
This package implements the partial Cox regression method proposed by Li and Gui (2004) and enables its application in an external data set. 
}
\details{
\tabular{ll}{
Package: \tab PartialCox\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-30\cr
License: \tab GPL-3\cr
}
Given a training data matrix and a survival object a partial Cox model can be trained. By default the number of evaluated latent variables is set to three but this can be controlled by the user. 
}
\author{
Miika Ahdesmaki

Maintainer: Miika Ahdesmaki <miika.ahdesmaki@live.fi>
}
\references{
Li, H. and J. Gui (2004), “Partial Cox regression analysis for high-dimensional microarray gene expression data,” Bioinformatics, 20(Suppl 1), i208–i215.
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
